Split up tests and remove f-strings in tests
